Increased takeover for club ammount. Implemented managerial changes based on board happiness throughout the season chosen from free agent managers. Implemented board happiness and morale features getting updated throughout the season. Clamped a lot of math stuff to stop exceptions and fixed a few random min bugs in math simulation.

+++ FootballChairman.cs
@@ -48,12 +48,12 @@
             this.Happiness = 100;
         }
 
-        public void UpdatePersonalFinance(int amount)
+        public void UpdatePersonalFinance(long amount)
         {
             this.PersonalFortune += amount;
         }
 
-        private ChairmanType GetChairmanType(int fortune)
+        private ChairmanType GetChairmanType(long fortune)
         {
             if(fortune > 1000000000)
             {
